+++ examples/test/constout/test_constout_bad_assign.pas
@@ -1,0 +1,7 @@
+program TestConstoutBadAssign;
+procedure P(const x: integer);
+begin
+    x := 5;
+end;
+begin
+end.