+++ examples/test/constout/test_constout_bad_both_var_const.pas
@@ -1,0 +1,6 @@
+program TestConstoutBadBothVarConst;
+procedure P(var const x: integer);
+begin
+end;
+begin
+end.