+++ examples/test/constout/test_constout_bad_forward_to_var.pas
@@ -1,0 +1,11 @@
+program TestConstoutBadForwardToVar;
+procedure Inner(var v: integer);
+begin
+    v := 1;
+end;
+procedure Outer(const c: integer);
+begin
+    Inner(c);
+end;
+begin
+end.