+++ examples/test/constout/test_constout_bad_incdec.pas
@@ -1,0 +1,7 @@
+program TestConstoutBadIncDec;
+procedure P(const x: integer);
+begin
+    inc(x);
+end;
+begin
+end.