+++ examples/test/constout/test_constout_bad_new.pas
@@ -1,0 +1,11 @@
+program TestConstoutBadNew;
+type
+    TFoo = class
+        v: integer;
+    end;
+procedure P(const c: TFoo);
+begin
+    new(c);
+end;
+begin
+end.