+++ examples/test/constout/test_constout_bad_proctype_scope.pas
@@ -1,0 +1,5 @@
+program TestConstoutBadProctypeScope;
+type
+    TProc = procedure(const x: integer);
+begin
+end.