+++ examples/test/constout/test_constout_const_basic.pas
@@ -1,0 +1,13 @@
+program TestConstoutConstBasic;
+var
+    n: integer;
+
+procedure ShowIt(const x: integer);
+begin
+    writeln(x);
+end;
+
+begin
+    n := 5;
+    ShowIt(n);
+end.