+++ examples/test/constout/test_constout_forward_const_to_const.pas
@@ -1,0 +1,18 @@
+program TestConstoutForwardConstToConst;
+var
+    n: integer;
+
+procedure Inner(const v: integer);
+begin
+    writeln(v);
+end;
+
+procedure Outer(const c: integer);
+begin
+    Inner(c);
+end;
+
+begin
+    n := 7;
+    Outer(n);
+end.