+++ examples/test/constout/test_constout_out_basic.pas
@@ -1,0 +1,13 @@
+program TestConstoutOutBasic;
+var
+    a: integer;
+
+procedure MakeIt(out y: integer);
+begin
+    y := 99;
+end;
+
+begin
+    MakeIt(a);
+    writeln(a);
+end.