+++ examples/test/constout/test_constout_warn_out_unassigned.pas
@@ -1,0 +1,12 @@
+program TestConstoutWarnOutUnassigned;
+var
+    a: integer;
+
+procedure Forgetful(out y: integer);
+begin
+    writeln('did nothing');
+end;
+
+begin
+    Forgetful(a);
+end.