+++ src/pascalc/lexer.c
@@ -231,6 +231,7 @@ void next_token(void) {
         else if (strcasecmp(token.text, "filesize") == 0) token.type = TOKEN_FILESIZE;
         else if (strcasecmp(token.text, "abstract") == 0) token.type = TOKEN_ABSTRACT;
         else if (strcasecmp(token.text, "sealed") == 0) token.type = TOKEN_SEALED;
+        else if (strcasecmp(token.text, "out") == 0) token.type = TOKEN_OUT;
         else token.type = TOKEN_IDENTIFIER;
         return;
     }